+++ Userland/Libraries/LibC/sys/types.h
@@ -75,8 +75,8 @@ typedef struct __pthread_mutexattr_t {
 } pthread_mutexattr_t;
 
 typedef struct __pthread_cond_t {
+    pthread_mutex_t* mutex;
     uint32_t value;
-    uint32_t previous;
     int clockid; // clockid_t
 } pthread_cond_t;
